Refactored Handlers

+++ sync/types.go
@@ -4,7 +4,7 @@ import "github.com/bits-and-blooms/bloom/v3"
 
 // Payload ...
 type Payload struct {
-	MissingElements []uint16           `json:"missingElements"`
+	MissingElements []uint32           `json:"missingElements"`
 	BF              *bloom.BloomFilter `json:"bf"`
 	Hash            uint64             `json:"hash"`
 }
